refactor: componentization of step2 in the service creation page
